+++ accession-commons-core/src/main/java/uk/ac/ebi/ampt2d/commons/accession/autoconfigure/BasicRestControllerAdviceConfiguration.java
@@ -23,7 +23,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
 import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
 import uk.ac.ebi.ampt2d.commons.accession.rest.BasicRestControllerAdvice;
-import uk.ac.ebi.ampt2d.commons.accession.rest.CollectionValidator;
+import uk.ac.ebi.ampt2d.commons.accession.rest.validation.CollectionValidator;
 
 /**
  * Basic configuration to inject a {@link BasicRestControllerAdvice} to return the appropriate response to the
