+++ accession-commons-core/src/main/java/uk/ac/ebi/ampt2d/commons/accession/rest/validation/CollectionValidator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  *
  */
-package uk.ac.ebi.ampt2d.commons.accession.rest;
+package uk.ac.ebi.ampt2d.commons.accession.rest.validation;
 
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
